Moved the UIViewController extension
