+++ FilmsReview/Scenes/Authentication/View/SignUpViewController.swift
@@ -159,16 +159,3 @@ final class SignUpViewController: UIViewController, SignUpVCProtocol {
     }
     
 }
-
-
-extension UIViewController {
-    func showErrorAlert(_ message: String, title: String = "Error") {
-        let alert = UIAlertController(
-            title: title,
-            message: message,
-            preferredStyle: .alert
-        )
-        alert.addAction(UIAlertAction(title: "OK", style: .default))
-        present(alert, animated: true)
-    }
-}
